requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

pandas==2.2.3
requests==2.32.3

pydantic==2.10.3
pydantic-settings==2.6.1

python-dotenv==1.0.1

pytest==8.3.4
httpx==0.28.1

scikit-learn==1.5.0